// src/app/shared/common/enums/activity.enums.ts
export enum ActivityEnum {
    Onboarding_Request = "101",
    Onboarding_Details = "102",
    MOU = "103",
    ProposalAndPI = "104",
    Payment = "105",
    SignOffAndDataHandover = "106",
    UC = "107",
    WorkOrderDetails = "201",
    EmployeeQualificationDetails = "301",
    EmployeeDetails = "401",
    ProfileDetails = "501",
    CounsellingDocuments = "601",
        
    /*
    Project_Documents = "108",
    */
}