// src/app/shared/model/app-project-activity.model.ts
export interface AppProjectActivityModel {
    id: Number;
   // projectId: Number;
    activityId: Number;
    status: String;
    submitTime: String;
    ipAddress: String;
    activityParentRefId:string;
   // activityName:string;

//     id: 0,
//     activityParentRefId: this.MOUdata.requestNo,
//     activityId: Number(ActivityEnum.MOU),
//     status: MdStatusEnum.Completed,
//     submitTime: "2023-10-19T12:53:17.277Z",
//     ipAddress: "",
// }
}